requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
sqlalchemy[asyncio]
asyncpg
alembic
pydantic
pydantic-settings
passlib
